Move the web server configuration to another file instead of the main class in order to allow the unit tests run.
